Renamed pacbio views and components to be more semantic.
