+++ src/components/pacbio/PacbioLibraryTubeItem.vue
