+++ src/views/pacbio/PacbioPlateIndex.vue
