+++ tests/unit/components/pacbio/PacbioRunPlateItem.spec.js
@@ -1,5 +1,5 @@
 import { mount, localVue, store } from '../../testHelper'
-import Plate from '@/components/pacbio/PacbioPlate'
+import Plate from '@/components/pacbio/PacbioRunPlateItem'
 import * as Run from '@/api/PacbioRun'
 import PlateMap from '@/config/PlateMap'
